fix: add modules
