+++ index.html
@@ -5,10 +5,14 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Maze</title>
-    <script src="js/constants.js" defer></script>
-    <script src="js/blanck-maze.js" defer></script>
-    <script src="js/maze.js" defer></script>
-    <link rel="stylesheet" href="style.css">
+    <!-- <script src="js/constants.js"></script>
+    <script src="js/sceen.js"></script>
+    <script src="js/backtrack-generator.js"></script>
+    <script src="js/index.js" defer></script> -->
+    <script type="module">
+      import './js/index.js';
+    </script>
+    <link rel="stylesheet" href="style.css" />
   </head>
   <body>
     <div class="table"></div>
