+++ package.json
@@ -1,0 +1,6 @@
+{
+  "name": "maze",
+  "version": "1.0.0",
+  "author": "Reda <inteex13@gmail.com>",
+  "license": "MIT"
+}
