+++ style.css
@@ -8,7 +8,7 @@
 }
 table {
   height: calc(100vh);
-  width: calc(100vh);
+  width: calc(100vw);
   border-spacing: 0;
 }
